backend/requirements.txt
Django==4.2.*
djangorestframework==3.14.*
djangorestframework-simplejwt==5.3.*
django-cors-headers==4.3.*
psycopg2-binary==2.9.*
bcrypt==4.1.*
python-decouple==3.8
django-filter==23.5
gunicorn==21.2.0
